requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
aiomysql
python-dotenv
requests
aiosqlite